+++ src/cherami/__init__.py
@@ -1,7 +1,7 @@
 from importlib.metadata import version
 
 __version__ = version("cherami")
-__codename__ = "Racing Homer"
+__codename__ = "Dutch Beauty"
 
 PIGEON = """\
      .-''-.