+++ src/cherami/pipelines/orange_box.py
@@ -9,7 +9,7 @@ from cherami.pipelines.pipeline import (
     PipelineContext,
     get_context_from_record,
 )
-from cherami.pipelines.worker import Worker
+from cherami.pipelines.worker import Worker, WorkerError
 
 logger = logging.getLogger(__name__)
 
@@ -160,6 +160,115 @@ class OrangeBoxPipeline(Pipeline):
 
 
 class OrangeBoxWorker(Worker):
+    def validate(self) -> None:
+        """
+        Overwrite method to check expected queues - priority and exchange.
+
+        Allows backwards compatbility (without running rerun and priority), but
+        will write a warning in the log.
+
+        If either the exchange OR queue suffix is given but the other is not,
+        this raises an error.
+
+        Acts as safety net check rather than user friendly descriptive UI.
+
+        Raises:
+            WorkerError - if any required checks fail.
+        """
+        # Listen
+        super().validate()
+
+        # Publish
+        if not self.publish_exchange or not self.publish_queue_suffix:
+            raise WorkerError(
+                "Orange box worker expects publish exchange and publish "
+                "queue suffix set - check worker config."
+            )
+
+        # Priority
+        if not self.priority_exchange and not self.priority_queue_suffix:
+            logger.warning(
+                "Orange Box Priority Message Queue not set, priority "
+                "messages will NOT be consumed."
+            )
+        if bool(self.priority_exchange) != bool(self.priority_queue_suffix):
+            raise WorkerError(
+                "For priority queue consumption, both the priority exchange "
+                "AND priority queue suffix must be set, check worker config. "
+            )
+
+        # Rerun
+        if not self.rerun_exchange and not self.rerun_queue_suffix:
+            logger.warning(
+                "Orange Box Rerun Message Queue not set, rerun "
+                "messages will NOT be consumed."
+            )
+        if bool(self.rerun_exchange) != bool(self.rerun_queue_suffix):
+            raise WorkerError(
+                "For rerun queue consumption, both the rerun exchange "
+                "AND rerun queue suffix must be set, check worker config. "
+            )
+
+    def get_message(self) -> Any | None:
+        """
+        Overwrites the default Worker method to handle priority and rerun
+        queues if provided in config.
+        Orange Box has three queues to consume from. The listening queue is the
+        main ingest queue, plus there is a low priority rerun queue for
+        messages that are being rerun through the entire pipeline, and finally
+        there is a high priority rerun queue local to orange box only, such
+        that everything downstream of chimera can be rerun.
+
+        Returns: one varys_client message object.
+        """
+        if self.priority_exchange:
+            priority_message: Any = self._varys_client.receive(
+                exchange=self.priority_exchange,
+                queue_suffix=self.priority_queue_suffix,
+                prefetch_count=1,
+                timeout=1,
+            )
+        else:
+            priority_message = None
+
+        main_message: Any = self._varys_client.receive(
+            exchange=self.listen_exchange,
+            queue_suffix=self.listen_queue_suffix,
+            prefetch_count=1,
+            timeout=1,
+        )
+        if self.rerun_exchange:
+            # low priority rerun queue
+            rerun_message: Any = self._varys_client.receive(
+                exchange=self.rerun_exchange,
+                queue_suffix=self.rerun_queue_suffix,
+                prefetch_count=1,
+                timeout=1,
+            )
+        else:
+            rerun_message = None
+
+        # Handle the message that is returned, nack any other messages if
+        # present
+        if priority_message:
+            message = priority_message
+            logger.info("Consuming from priority queue.")
+            if rerun_message:
+                self._varys_client.nack_message(rerun_message)
+            if main_message:
+                self._varys_client.nack_message(main_message)
+        elif main_message:
+            message = main_message
+            logger.info("Consuming from main queue.")
+            if rerun_message:
+                self._varys_client.nack_message(rerun_message)
+        elif rerun_message:
+            message = rerun_message
+            logger.info("Consuming from rerun queue.")
+        else:
+            message = None
+        return message
+
     def on_success(self, message: Any, context: PipelineContext) -> None:
         """Handle successful orange box pipeline completions.